Move bin code into engine.
Why was this in the binary crate anyway?
