+++ crates/oneiros-engine/Cargo.toml
@@ -35,6 +35,7 @@ sha2.workspace = true
 thiserror.workspace = true
 tokio.workspace = true
 tokio-stream.workspace = true
+tracing-subscriber.workspace = true
 toml.workspace = true
 tower.workspace = true
 uuid.workspace = true
