+++ crates/oneiros/Cargo.toml
@@ -14,8 +14,5 @@ name = "oneiros"
 path = "src/main.rs"
 
 [dependencies]
-anstream.workspace = true
 oneiros-engine.workspace = true
-serde_json.workspace = true
 tokio.workspace = true
-tracing-subscriber.workspace = true
